+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,38 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.transaction.support.TransactionTemplate;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionTemplate transactionTemplate;
+    User findById(long id);
 
-    public UserService(UserDao userDao, UserHistoryDao userHistoryDao, TransactionTemplate transactionTemplate) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = transactionTemplate;
-    }
+    void insert(User user);
 
-    public User findById(long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(long id, String newPassword, String createBy) {
-        transactionTemplate.execute(() -> {
-            var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            return null;
-        });
-    }
+    void changePassword(long id, String newPassword, String createBy);
 }
